+++ geodataset/tools.py
@@ -1,5 +1,5 @@
 from geodataset.utils import BadAreaDefinition
-from geodataset.geodataset import GeoDataset
+from geodataset.geo_dataset import GeoDataset
 from geodataset.nextsim_moorings import Moorings
 
 def open_netcdf(file_address):
